Remove uncompressed blocks from deflate lists.
Currently, we are including all the uncompressed blocks in the deflate
buffer. But that is totally unneeded since that is just an overhead and there is
no special magic puffin conversions happening. This patch removes the
uncompressed blocks in FindDeflateSublocks() function. The puffdiff and other
use cases still have to process uncompressed blocks.

Bug: 80148425
Test: unitest
Test: ran on a corpus
Change-Id: I2ccf1200dd6ebb2d244d992054c1f2c4647c4c2d