Prevent fuzzer from going out-of-memory on valid but outragous input data
Currently, the puffpatch fuzzer fails once in a while with out-of-memory or
segfaults that are caused by valid input data but such that the requested puff
and deflate buffers are very large. To fix this we first decode the header and
make sure the deflate and puff buffer sizes do not excede some limits. Then, we
go ahead with calling the puffpatch.

Bug: crbug.com/848058
Test: USE="fuzzer asan" FEATURES=test emerge-amd64-generic puffin

Change-Id: Iedf9b7cf6988322364bbef944296dfc87c099d7d