Use Puffer (instead of zlib library) to find deflate locations
Currently, we are dependent on zlib to find deflate locations in gzip and zip
archives. This means we have to pass a deflate stream one time to zlib library
and two times to puffer (to find deflate subblocks locations and puff
locations). This patch changes the puffer function to terminate the call when
the final block is reached. Hence, it can be used easily to identify the
location of deflate subblocks by only one pass (plus one pass for puff
locations). This will reduce the delta payload generation time.

Bug: 80148379
Test: unittests
Test: test_corpus.py
Test: measure_patch_size.py

Change-Id: I4e7e6ed1edc2d5636124600d41f776332cda570a