Don't return false if there was no compressed deflate in a stream
Currently, LocateDeflatesInGzip() fails if the stream either does not
have any deflates, or the deflates are all uncompressed. This can happen
for example if there are gzip files in a squashfs image compressed with
gzip. With high probability, the gzip won't double compress the
compressed files because that might actually increase the final size. So
we should not fail on this function if there was no deflates
indentified same as LocateDeflatesInZipArchive().

Bug: crbug.com/1050869
Test: cros_generate_update_payload --image gs://chromeos-releases/dev-channel/eve-kvm/12871.1.0/dlc/pita/package/dlc.img --src_image gs://chromeos-releases/dev-channel/eve-kvm/12861.0.0/dlc/pita/package/dlc.img --output delta.bin --debug

Change-Id: Ic770d8cb01e14306e7e49cdaadf09b5900e49d17